CAS-347: Add /add-fungible-token
