+++ frontend/packages/client/src/hooks/index.js
@@ -21,3 +21,4 @@ export { default as useCommunityUsers } from './useCommunityUsers';
 export { default as useCommunityCategory } from './useCommunityCategory';
 export { default as useBeforeUnload } from './useBeforeUnload';
 export { default as useStarAnimation } from './useStarAnimation';
+export { default as useAddFungibleToken } from './useAddFungibleToken';
