+++ frontend/packages/client/src/hooks/useVotingStrategies.js
@@ -23,40 +23,8 @@ export default function useVotingStrategies() {
     getVotingStrategies();
   }, [getVotingStrategies]);
 
-  const addFungibleToken = useCallback(
-    async (addr, name) => {
-      const url = `${process.env.REACT_APP_BACK_END_SERVER_API}/add-fungible-token`;
-      try {
-        const fetchOptions = {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            addr,
-            name,
-          }),
-        };
-        dispatch({ type: 'PROCESSING' });
-        const response = await fetch(url, fetchOptions);
-        const json = await checkResponse(response);
-        dispatch({
-          type: 'SUCCESS',
-          payload: json,
-        });
-        return json;
-      } catch (err) {
-        notifyError(err, url);
-        dispatch({ type: 'ERROR', payload: { errorData: err.message } });
-        return { error: err.message };
-      }
-    },
-    [dispatch, notifyError]
-  );
-
   return {
     ...state,
     getVotingStrategies,
-    addFungibleToken,
   };
 }
